Support & optimizations

+++ assets/app/admin/business/review.component.ts
@@ -12,6 +12,8 @@ import { BusinessService } from './business.service';
 
 export class ReviewComponent implements OnInit {
   private businesses: Business[] = [];
+  private loading: boolean = false;
+
   constructor(private businessService: BusinessService) { }
   ngOnInit() {
     this.businessService.getBusinesses().subscribe(businesses => {
@@ -19,16 +21,38 @@ export class ReviewComponent implements OnInit {
     });
   }
 
+  isLoading() {
+    return this.loading;
+  }
+
   deleteBusiness(businessId: string) {
-    this.businessService.deleteBusiness(businessId).subscribe(msg => {
-      alert(msg);
-      location.reload();
+    bootbox.confirm({
+      title: "Delete Business?",
+      message: "Are you sure you want to delete this business? This cannot be undone.",
+      buttons: {
+        cancel: {
+          label: '<i class="fa fa-times"></i> Cancel'
+        },
+        confirm: {
+          label: '<i class="fa fa-check"></i> Delete'
+        }
+      },
+      callback: (result) => {
+        if (result) {
+          this.loading = true;
+          this.businessService.deleteBusiness(businessId).subscribe(msg => {
+            bootbox.alert(msg, () => {
+              location.reload();
+            });
+          });
+        }
+      }
     });
   }
 
   viewBusiness(businessId: string) {
     //To be adjusted for the user profile route.
-    alert('To be redirected to profile');
+    bootbox.alert('To be redirected to profile');
     // this.router.navigate(['admin', userId]);
   }
 };
